# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Altsuite backend API documentation",
        "title": "Altsuite API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/docker/containers": {
            "get": {
                "description": "List all Docker containers running on the system",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Docker"
                ],
                "summary": "List Docker containers",
                "responses": {
                    "200": {
                        "description": "Docker containers output",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Docker not installed or error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns service health and timestamp",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HealthResponse"
                        }
                    }
                }
            }
        },
        "/metrics/current": {
            "get": {
                "description": "Returns the most recent system metrics snapshot (CPU, Memory, Network, Disk usage)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Get current metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collectors.MetricsPoint"
                        }
                    },
                    "503": {
                        "description": "No metrics available yet",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/metrics/history": {
            "get": {
                "description": "Returns historical metrics for a specified time range with automatic downsampling for longer periods",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Get historical metrics",
                "parameters": [
                    {
                        "type": "string",
                        "default": "minute",
                        "description": "Time range: minute (default), hour, day, week, or month",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.historyResponse"
                        }
                    }
                }
            }
        },
        "/packages": {
            "get": {
                "description": "Return list of installed packages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "List packages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PackageListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/packages/install": {
            "post": {
                "description": "Install one or more supported packages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Install packages",
                "parameters": [
                    {
                        "description": "Package names",
                        "name": "packages",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/packages/{name}": {
            "get": {
                "description": "Retrieve detailed information about a specific package",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Packages"
                ],
                "summary": "Get package information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package information",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/action": {
            "post": {
                "description": "Start/stop/restart/enable/disable a system service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Perform a service action",
                "parameters": [
                    {
                        "description": "Service action request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ServiceActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ServiceStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/{name}/status": {
            "get": {
                "description": "Return running status for a given service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get service status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ServiceStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "collectors.MetricsPoint": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "number"
                },
                "disk": {
                    "type": "number"
                },
                "memory": {
                    "type": "number"
                },
                "network": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "controllers.historyResponse": {
            "type": "object",
            "properties": {
                "metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/collectors.MetricsPoint"
                    }
                },
                "range": {
                    "type": "string"
                }
            }
        },
        "main.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "main.PackageListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "packages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.SupportedApp"
                    }
                }
            }
        },
        "main.ServiceActionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "start, stop, restart, enable, disable",
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "main.ServiceStatusResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "is_running": {
                    "type": "boolean"
                },
                "output": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "main.Spec": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                }
            }
        },
        "main.SupportedApp": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "installed": {
                    "type": "boolean"
                },
                "monthlySavings": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "replaces": {
                    "type": "string"
                },
                "requiredSpecs": {
                    "$ref": "#/definitions/main.Spec"
                }
            }
        }
    }
}