# api/main.go
// @title Altsuite API
// @version 1.0
// @description Altsuite backend API documentation
// @termsOfService http://swagger.io/terms/
// @contact.name API Support
// @contact.url http://www.example.com/support
// @contact.email support@example.com
// @license.name MIT
// @license.url https://opensource.org/licenses/MIT
// @host localhost:8080
// @BasePath /api

package main

import (
	"context"
	"encoding/json"
	"log"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/gorilla/mux"
	_ "github.com/smalex/altsuite/docs"
	httpSwagger "github.com/swaggo/http-swagger"

	"github.com/smalex/altsuite/collectors"
	"github.com/smalex/altsuite/routes"
)

type HealthResponse struct {
	Status    string    `json:"status"`
	Timestamp time.Time `json:"timestamp"`
}

type ServiceStatusResponse struct {
	ServiceName string `json:"service_name"`
	IsRunning   bool   `json:"is_running"`
	Output      string `json:"output,omitempty"`
	Error       string `json:"error,omitempty"`
}

type ServiceActionRequest struct {
	ServiceName string `json:"service_name"`
	Action      string `json:"action"` // start, stop, restart, enable, disable
}

type PackageListResponse struct {
	Packages []SupportedApp `json:"packages"`
	Count    int            `json:"count"`
}

const shutdownTimeout = 5 * time.Second

var privOps *PrivilegedOps

func main() {
	// Initialize privileged operations handler
	privOps = NewPrivilegedOps()

	// Create a cancellable context for the metrics collector; it will be
	// cancelled during graceful shutdown so the background goroutine exits.
	ctx, cancel := context.WithCancel(context.Background())
	defer cancel()

	// Start metrics collector
	metricsCollector := collectors.NewMetricsCollector()
	metricsCollector.Start(ctx)

	r := mux.NewRouter()

	// API routes
	api := r.PathPrefix("/api").Subrouter()
	api.HandleFunc("/health", healthHandler).Methods("GET")

	// System metrics endpoints
	routes.RegisterMetricsRoutes(api, metricsCollector)

	// Service management endpoints
	api.HandleFunc("/services/{name}/status", getServiceStatusHandler).Methods("GET")
	api.HandleFunc("/services/action", serviceActionHandler).Methods("POST")

	// Package management endpoints
	api.HandleFunc("/packages", listPackagesHandler).Methods("GET")
	api.HandleFunc("/packages/{name}", getPackageInfoHandler).Methods("GET")
	api.HandleFunc("/packages/install", installPackageHandler).Methods("POST")

	// Docker endpoints (if Docker is installed)
	api.HandleFunc("/docker/containers", listDockerContainersHandler).Methods("GET")

	// Serve swagger UI (generated by `swag init` into ./docs)
	// The swagger JSON will be expected at /docs/swagger.json
	r.PathPrefix("/swagger/").Handler(httpSwagger.WrapHandler)

	// Serve frontend static files
	frontendDir := "/opt/altsuite/frontend"
	r.PathPrefix("/").Handler(http.FileServer(http.Dir(frontendDir)))

	// CORS middleware for development
	r.Use(corsMiddleware)

	// Start server
	port := ":8080"
	srv := &http.Server{
		Addr:    port,
		Handler: r,
	}

	// Listen for OS signals to trigger graceful shutdown.
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)

	go func() {
		log.Printf("Server starting on port %s", port)
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatal(err)
		}
	}()

	<-quit
	log.Println("Server shutting down...")

	// Cancel the metrics collector context first.
	cancel()

	// Give in-flight requests up to 5 seconds to complete.
	shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), shutdownTimeout)
	defer shutdownCancel()
	if err := srv.Shutdown(shutdownCtx); err != nil {
		log.Printf("Server forced to shutdown: %v", err)
	}
}

// healthHandler godoc
// @Summary Health check
// @Description Returns service health and timestamp
// @Tags Health
// @Produce json
// @Success 200 {object} HealthResponse
// @Router /health [get]
func healthHandler(w http.ResponseWriter, r *http.Request) {
	response := HealthResponse{
		Status:    "healthy",
		Timestamp: time.Now(),
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// getServiceStatusHandler godoc
// @Summary Get service status
// @Description Return running status for a given service
// @Tags Services
// @Produce json
// @Param name path string true "Service name"
// @Success 200 {object} ServiceStatusResponse
// @Failure 404 {object} map[string]string
// @Router /services/{name}/status [get]
func getServiceStatusHandler(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	serviceName := vars["name"]

	isRunning, err := privOps.GetServiceStatus(serviceName)

	response := ServiceStatusResponse{
		ServiceName: serviceName,
		IsRunning:   isRunning,
	}

	if err != nil {
		response.Error = err.Error()
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// serviceActionHandler godoc
// @Summary Perform a service action
// @Description Start/stop/restart/enable/disable a system service
// @Tags Services
// @Accept json
// @Produce json
// @Param request body ServiceActionRequest true "Service action request"
// @Success 200 {object} ServiceStatusResponse
// @Failure 400 {object} map[string]string
// @Router /services/action [post]
func serviceActionHandler(w http.ResponseWriter, r *http.Request) {
	var req ServiceActionRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	var operation SystemctlOperation
	switch req.Action {
	case "start":
		operation = SystemctlStart
	case "stop":
		operation = SystemctlStop
	case "restart":
		operation = SystemctlRestart
	case "enable":
		operation = SystemctlEnable
	case "disable":
		operation = SystemctlDisable
	default:
		http.Error(w, "Invalid action. Use: start, stop, restart, enable, or disable", http.StatusBadRequest)
		return
	}

	output, err := privOps.SystemctlCommand(operation, req.ServiceName)

	response := ServiceStatusResponse{
		ServiceName: req.ServiceName,
		Output:      output,
	}

	if err != nil {
		response.Error = err.Error()
		w.WriteHeader(http.StatusInternalServerError)
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// listPackagesHandler godoc
// @Summary List packages
// @Description Return list of installed packages
// @Tags Packages
// @Produce json
// @Success 200 {object} PackageListResponse
// @Failure 500 {object} map[string]string
// @Router /packages [get]
func listPackagesHandler(w http.ResponseWriter, r *http.Request) {
	packages, err := privOps.ListInstalledPackages()
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	response := PackageListResponse{
		Packages: packages,
		Count:    len(packages),
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// Handler for getting package info
// getPackageInfoHandler godoc
// @Summary Get package information
// @Description Retrieve detailed information about a specific package
// @Tags Packages
// @Produce plain
// @Param name path string true "Package name"
// @Success 200 {string} string "Package information"
// @Failure 404 {object} map[string]string
// @Router /packages/{name} [get]
func getPackageInfoHandler(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	packageName := vars["name"]

	info, err := privOps.GetPackageInfo(packageName)

	if err != nil {
		http.Error(w, err.Error(), http.StatusNotFound)
		return
	}

	w.Header().Set("Content-Type", "text/plain")
	w.Write([]byte(info))
}

// installPackageHandler godoc
// @Summary Install packages
// @Description Install one or more supported packages
// @Tags Packages
// @Accept json
// @Produce json
// @Param packages body []string true "Package names"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /packages/install [post]
func installPackageHandler(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Packages []string `json:"packages"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if len(req.Packages) == 0 {
		http.Error(w, "No packages specified", http.StatusBadRequest)
		return
	}

	output, err := privOps.PackageCommand(PackageInstall, req.Packages...)

	response := map[string]interface{}{
		"output": output,
	}

	if err != nil {
		response["error"] = err.Error()
		w.WriteHeader(http.StatusInternalServerError)
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// Handler for listing Docker containers
// listDockerContainersHandler godoc
// @Summary List Docker containers
// @Description List all Docker containers running on the system
// @Tags Docker
// @Produce plain
// @Success 200 {string} string "Docker containers output"
// @Failure 500 {object} map[string]string "Docker not installed or error"
// @Router /docker/containers [get]
func listDockerContainersHandler(w http.ResponseWriter, r *http.Request) {
	output, err := privOps.ListDockerContainers()

	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "text/plain")
	w.Write([]byte(output))
}

func corsMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Access-Control-Allow-Origin", "*")
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")

		if r.Method == "OPTIONS" {
			w.WriteHeader(http.StatusOK)
			return
		}

		next.ServeHTTP(w, r)
	})
}